pyproject.toml
[tool.poetry]
name = "parrot"
version = "3.0.0"
description = ""
authors = ["garlicOS® <14845347+Grosserly@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
"discord.py" = {extras = ["voice"], version = "^1.5.1"}
jishaku = "^1.19.2"
markovify = "^0.8.3"
ujson = "^4.0.1"
redis = "^3.5.3"
Pillow = "^8.3.1"
asyncstdlib = "^3.9.2"
aioredis = {extras = ["hiredis"], version = "^2.0.0"}

[tool.poetry.dev-dependencies]
mypy = "^0.790"
"discord.py-stubs" = "^1.5.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
